+++ src/GUI/textbox.py
@@ -154,8 +154,7 @@ class ScrollText(tk.Frame):
 
         if file_menu.return_file() == ".cpp" or file_menu.return_file() == ".py":
             self.highlight_syntax()
-            if file_menu.return_file() == ".cpp":
-                self.suggestions.update_suggestions()
+            self.suggestions.update_suggestions()
         self.numberLines.redraw()
         self.statusbar.update_stats(self.text)
         font_size = check.get_config_value("zoom")