+++ src/MainMenu/themes.py
@@ -94,6 +94,7 @@ def use_theme(init, menu_bar, cascade1, cascade2, cascade3, cascade4, cascade5, 
                 widget.suggestions.tags_text.configure(bg=config['suggestions']['bg'], fg=config['suggestions']['fg'],
                                              selectbackground=config['suggestions']['selectbackground'],
                                              selectforeground=config['suggestions']['selectforeground'])
+                widget.suggestions.scrollbar.configure(fg_color=config['suggestions']['bg'],button_color=config['suggestions']['fg'],button_hover_color=config['suggestions']['selectbackground'])
                 widget.terminal.configure(fg_color=config['terminal']['fg_color'], bg_color=config['terminal']['bg_color'])
                 widget.terminal.textbox.configure(fg_color=config['terminal']['textbox_color'], text_color=config['terminal']['text_color'])
             elif widget_name == 'tree':